refactor(home): move header and delete from router file
