--- src/routes/components/Header/index.ts
@@ -1,2 +1,0 @@
-export * from './Header.home';
-export * from './Header.detail';
