+++ src/screens/Home/components/Header/Header.styles.ts
@@ -1,5 +1,5 @@
 import {StyleSheet} from 'react-native';
-import { COLORS } from '../../../assets/colors';
+import {COLORS} from '../../../../assets/colors';
 
 export default StyleSheet.create({
   container: {
