+++ src/screens/Home/components/index.ts
@@ -1,3 +1,4 @@
 export * from './AllButton';
 export * from './PointsCard';
 export * from './ProductsContainer';
+export * from './Header';
